+++ R/predict.R
@@ -345,17 +345,17 @@ predict.sdmTMB <- function(object, newdata = object$data,
       cli_abort(c("There is at least one NA value in the time column.",
         "Please remove it."))
 
-    # newdata$sdm_orig_id <- seq(1, nrow(newdata))
-    # fake_newdata <- unique(newdata[,xy_cols])
-    # fake_newdata[["sdm_spatial_id"]] <- seq(1, nrow(fake_newdata)) - 1L
-    newdata$sdm_spatial_id <- seq(1, nrow(newdata)) # FIXME doing nothing now
-    #
-    # newdata <- base::merge(newdata, fake_newdata, by = xy_cols,
-    #   all.x = TRUE, all.y = FALSE)
-    # newdata <- newdata[order(newdata$sdm_orig_id),, drop = FALSE]
+    newdata$sdm_orig_id <- seq(1, nrow(newdata))
+    unique_newdata <- unique(newdata[,xy_cols])
+    unique_newdata[["sdm_spatial_id"]] <- seq(1, nrow(unique_newdata)) - 1L
+    # newdata$sdm_spatial_id <- seq(1, nrow(newdata)) - 1L
+    # FIXME make faster:
+    newdata <- base::merge(newdata, unique_newdata, by = xy_cols,
+      all.x = TRUE, all.y = FALSE)
+    newdata <- newdata[order(newdata$sdm_orig_id),, drop = FALSE]
 
     proj_mesh <- INLA::inla.spde.make.A(object$spde$mesh,
-      loc = as.matrix(newdata[,xy_cols, drop = FALSE]))
+      loc = as.matrix(unique_newdata[,xy_cols, drop = FALSE]))
 
     if (length(object$formula) == 1L) {
       # this formula has breakpt() etc. in it:
@@ -420,7 +420,7 @@ predict.sdmTMB <- function(object, newdata = object$data,
     tmb_data$exclude_RE <- exclude_RE
     # tmb_data$calc_index_totals <- as.integer(!se_fit)
     # tmb_data$calc_cog <- as.integer(!se_fit)
-    tmb_data$proj_spatial_index <- newdata$sdm_spatial_id - 1L
+    tmb_data$proj_spatial_index <- newdata$sdm_spatial_id
     tmb_data$proj_Zs <- sm$Zs
     tmb_data$proj_Xs <- sm$Xs